More code review comments
